do fetching at top level + add some new states + minor changes
