+++ client/src/index.css
@@ -50,3 +50,16 @@ code {
     opacity: 1;
   }
 }
+
+.fadesIn {
+  animation: fadeIn 0.5s 1;
+}
+
+@keyframes fadeIn {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
