+++ client/src/redux/handleAsync.js
@@ -3,10 +3,11 @@ import { REQUEST_STATE } from './states'
 export const handleAsyncAction = (builder, asyncAction, handlers) => {
   builder
     .addCase(asyncAction.pending, (state) => {
-      state.status = REQUEST_STATE.PENDING
+      handlers.pending
+        ? handlers.pending(state)
+        : (state.status = REQUEST_STATE.PENDING)
     })
     .addCase(asyncAction.fulfilled, (state, action) => {
-      state.status = REQUEST_STATE.FULFILLED
       handlers.fulfilled(state, action)
     })
     .addCase(asyncAction.rejected, (state, action) => {
