+++ client/src/redux/states.js
@@ -3,4 +3,7 @@ export const REQUEST_STATE = {
   PENDING: 'PENDING',
   FULFILLED: 'FULFILLED',
   REJECTED: 'REJECTED',
+
+  WRITING: 'WRITING',
+  READING: 'READING',
 }
